+++ src/renderer/ScreenController.tsx
@@ -20,6 +20,7 @@ import { debouncedUpdateBadgeCounter } from './system-integration/badge-counter'
 import { updateDeviceChats } from './deviceMessages'
 import { runtime } from './runtime'
 import { DcEventType } from '@deltachat/jsonrpc-client'
+import WebxdcSaveToChatDialog from './components/dialogs/WebxdcSendToChatDialog'
 
 const log = getLogger('renderer/ScreenController')
 
@@ -161,6 +162,12 @@ export default class ScreenController extends Component {
     }
 
     runtime.onOpenQrUrl = processOpenQrUrl
+    runtime.onWebxcSendToChat = (file, text) => {
+      ;(this.openDialog as OpenDialogFunctionType)(WebxdcSaveToChatDialog, {
+        messageText: text,
+        file,
+      })
+    }
 
     this.startup().then(() => {
       runtime.emitUIFullyReady()