+++ static/webxdc.d.ts
@@ -9,6 +9,9 @@ type SendingStatusUpdate<T> = {
    * usually only one line of text is shown,
    * use this option sparingly to not spam the chat. */
   info?: string
+  /** optional, if the Webxdc creates a document, you can set this to the name of the document;
+   * do not set if the Webxdc does not create a document */
+  document?: string
   /** optional, short text, shown beside the icon;
    * it is recommended to use some aggregated value,
    * eg. "8 votes", "Highscore: 123" */
@@ -24,9 +27,40 @@ type ReceivedStatusUpdate<T> = {
   max_serial: number
   /** optional, short, informational message. */
   info?: string
+  /** optional, if the Webxdc creates a document, this is the name of the document;
+   * not set if the Webxdc does not create a document */
+  document?: string
   /** optional, short text, shown beside the webxdc's icon. */
   summary?: string
 }
+
+type XDCFile = {
+  /** name of the file */
+  name: string
+} & (
+  | {
+      /** blob, also accepts types that inherit Blob, like File */
+      blob: Blob
+    }
+  | {
+      /** base64 encoded file data */
+      base64: string
+    }
+  | {
+      /** text for textfile, will be encoded as utf8 */
+      plainText: string
+    }
+)
+
+type sendOptions =
+  | {
+      file: XDCFile
+      text?: string
+    }
+  | {
+      file?: XDCFile
+      text: string
+    }
 
 interface Webxdc<T> {
   /** Returns the peer's own address.
@@ -39,21 +73,30 @@ interface Webxdc<T> {
    * set a listener for new status updates.
    * The "serial" specifies the last serial that you know about (defaults to 0).
    * Note that own status updates, that you send with {@link sendUpdate}, also trigger this method
+   * @returns promise that resolves when the listener has processed all the update messages known at the time when `setUpdateListener` was called.
    * */
   setUpdateListener(
     cb: (statusUpdate: ReceivedStatusUpdate<T>) => void,
-    serial: number
+    serial?: number
   ): Promise<void>
   /**
-   * WARNING! This function is deprecated, see setUpdateListener().
+   * @deprecated See {@link setUpdateListener|`setUpdateListener()`}.
    */
   getAllUpdates(): Promise<ReceivedStatusUpdate<T>[]>
   /**
-   * Webxdcs are usually shared in a chat and run independently on each peer. To get a shared status, the peers use sendUpdate() to send updates to each other.
+   * Webxdc are usually shared in a chat and run independently on each peer. To get a shared status, the peers use sendUpdate() to send updates to each other.
    * @param update status update to send
    * @param description short, human-readable description what this update is about. this is shown eg. as a fallback text in an email program.
    */
   sendUpdate(update: SendingStatusUpdate<T>, description: string): void
+  /**
+   * Send a message with file, text or both to a chat.
+   * Asks user to what Chat to send the message to.
+   * Always exits the xdc, please save your app state before calling this function
+   * @param content
+   * @returns returns a promise that never resolves (because the xdc closes), but is rejected on error.
+   */
+  sendToChat(content: sendOptions): Promise<void>
 }
 
 ////////// ANCHOR: global
@@ -64,4 +107,4 @@ declare global {
 }
 ////////// ANCHOR_END: global
 
-export { SendingStatusUpdate, ReceivedStatusUpdate, Webxdc }
+export { SendingStatusUpdate, ReceivedStatusUpdate, Webxdc, XDCFile }